pyproject.toml
[tool.poetry]
name = "с05-hw27-django_app"
version = "0.1.0"
description = ""
authors = ["kpaveliev <kpaveliev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
